Reviewed-on: https://git.smartdata.solutions/smartdata/smartskills/pulls/45

+++ src/server/routers/positionsRouter.ts
@@ -1,3 +1,4 @@
+import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 
 import { procedure, router } from "../trpc";
@@ -18,6 +19,30 @@ export const positionsRouter = router({
   deletePosition: procedure
     .input(z.object({ positionId: z.number() }))
     .mutation(async ({ ctx, input }) => {
+      const positonTitle = await ctx.prisma.positions.findUnique({
+        where: {
+          id: input.positionId,
+        },
+      });
+      const isPositionUsed = await ctx.prisma.user.findFirst({
+        where: {
+          positions: {
+            some: {
+              title: positonTitle?.title,
+            },
+          },
+        },
+        include: {
+          positions: true,
+          skills: true,
+        },
+      });
+      if (isPositionUsed) {
+        throw new TRPCError({
+          code: "CONFLICT",
+          message: "The position is in use and can't be deleted.",
+        });
+      }
       return await ctx.prisma.positions.delete({
         where: {
           id: input.positionId,
